pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vyuu-copilot-v2"
version = "0.1.0"
description = "LangGraph-based intent orchestration system"
readme = "README.md"
authors = ["Vyuu Team"]
homepage = "https://github.com/vyuu/vyuu-copilot-v2"
repository = "https://github.com/vyuu/vyuu-copilot-v2"
packages = [
    { include = "vyuu_copilot_v2", from = "src" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langgraph = ">=0.1.0"
langchain = ">=0.1.0"
supabase = ">=2.0.0"
pydantic = ">=2.0.0"
fastapi = ">=0.100.0"
uvicorn = {extras = ["standard"], version = ">=0.20.0"}
pytest = ">=7.0.0"
python-dotenv = ">=1.0.0"
asyncpg = ">=0.28.0"
tenacity = ">=8.0.0"
pyjwt = ">=2.8.0"
gotrue = ">=1.0.0"
openai = ">=1.0.0"
pyyaml = ">=6.0.0"

[tool.poetry.group.dev.dependencies]
black = ">=23.0.0"
isort = ">=5.12.0"
flake8 = ">=6.0.0"
mypy = ">=1.0.0"
pre-commit = ">=3.0.0"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests requiring external dependencies",
]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]